% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_errors.R
\name{bind_errors}
\alias{bind_errors}
\title{Bind errors stored as JSON files}
\usage{
bind_errors(data_path, verbose = TRUE)
}
\arguments{
\item{data_path}{string, file path to directory of stored tweets data saved as errors_\emph{id}.json}

\item{verbose}{If \code{FALSE}, messages are suppressed}
}
\value{
a data.frame containing error information
}
\description{
This function binds the errors stored as JSON files. Errors are only returned if requested in the retrieving function.
}
\details{
By default, \code{bind_errors} binds into a data frame containing tweets (from errors_\emph{id}.json files).
}
\examples{
\dontrun{
# retrieve data with errors and store them locally as .json
get_all_tweets(query = "BLM", 
               start_tweets = "2020-01-01T00:00:00Z", 
               end_tweets = "2020-01-05T00:00:00Z", 
               bearer_token = bearer_token, 
               data_path = "data/",
               n = 500, 
               bind_tweets = F,
               errors = T)

# bind json files in the directory "data" into a data frame containing errors
bind_errors(data_path = "data/")
}
}
