% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{bind_user_jsons}
\alias{bind_user_jsons}
\title{Bind user information stored as JSON files
`r lifecycle::badge("deprecated")}
\usage{
bind_user_jsons(data_path, verbose = TRUE)
}
\arguments{
\item{data_path}{string, file path to directory of stored tweets data saved as users_\emph{id}.json}

\item{verbose}{If \code{FALSE}, messages are suppressed}
}
\value{
a data.frame
}
\description{
Bind user information stored as JSON files
`r lifecycle::badge("deprecated")
}
\examples{
\dontrun{
bind_user_jsons("data_path = "data/"")
}
}
\keyword{internal}
