% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compliance.R
\name{create_compliance_job}
\alias{create_compliance_job}
\title{Create Compliance Job}
\usage{
create_compliance_job(
  file,
  type = "tweets",
  bearer_token = get_bearer(),
  verbose = TRUE
)
}
\arguments{
\item{file}{a text file, each line contains a Tweet ID or user ID.}

\item{type}{the type of the job, whether "tweets" or "users".}

\item{bearer_token}{string, bearer token}

\item{verbose}{If \code{FALSE}, query progress messages are suppressed}
}
\value{
the job ID (invisibly)
}
\description{
This function creates a new compliance job and upload the Tweet IDs or user IDs.
}
\examples{
\dontrun{
create_compliance_job("tweetids.txt", "tweets")
}
}
