% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bearer.R
\name{get_bearer}
\alias{get_bearer}
\title{Manage your bearer token}
\usage{
get_bearer()
}
\value{
string represents your bearer token, if it the environmental variable "TWITTER_BEARER" has been preset.
}
\description{
This function attempts to retrieve your bearer token from the environmental variable "TWITTER_BEARER". The easiest way to setup this environmental variable is to use \code{usethis::edit_r_environ()} and insert this line to your \code{.Renviron} file: TWITTER_BEARER=YOURTOKENHERE. Replace YOURTOKEN here with your own token.
}
