% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compliance.R
\name{get_compliance_result}
\alias{get_compliance_result}
\title{Get Compliance Result}
\usage{
get_compliance_result(id, bearer_token = get_bearer())
}
\arguments{
\item{id}{string, the id of the compliance job}

\item{bearer_token}{string, bearer token}
}
\value{
a data frame
}
\description{
This function retrieves the information for a single compliance job.
}
\examples{
\dontrun{
get_compliance_result("1460077048991555585")
}
}
