% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geo_tweetsv2.R
\name{get_geo_tweets}
\alias{get_geo_tweets}
\title{Get tweets for query containing geo information}
\usage{
get_geo_tweets(
  query,
  start_tweets,
  end_tweets,
  bearer_token,
  file = NULL,
  data_path = NULL
)
}
\arguments{
\item{query}{}

\item{start_tweets}{}

\item{end_tweets}{}

\item{bearer_token}{}

\item{file}{}

\item{data_path}{}
}
\value{

}
\description{
This function loops through specified strings or hashtags and collects tweets containing the strings or hashtags between specified date ranges that also contain Tweet-specific geolocation data provided by the Twitter user. This can be either a location in the form of a Twitter place, with the corresponding display name, geo polygon, and other fields, or in rare cases, a geo lat-long coordinate. Note: Operators matching on place (Tweet geo) will only include matches from original tweets. Tweet-level data is stored in a data/ path as a series of JSONs beginning "data_"; User-level data is stored as a series of JSONs beginning "users_". If a filename is supplied, this function will save the result as a RDS file, otherwise, it will return the results as a dataframe.
}
