% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hashtag_tweetsv2.R
\name{get_hashtag_tweets}
\alias{get_hashtag_tweets}
\title{Get tweets from user}
\usage{
get_hashtag_tweets(
  query,
  start_tweets,
  end_tweets,
  bearer_token,
  file = NULL,
  data_path = NULL
)
}
\arguments{
\item{query}{string, search query, use "+" to separate query terms.}

\item{start_tweets}{string, starting date}

\item{end_tweets}{string, ending date}

\item{bearer_token}{string, bearer token}

\item{file}{string, name of the resulting RDS file. Will return a dataframe if not supplied}

\item{data_path}{string, if supplied, fetched data will be saved to the designated path as jsons}
}
\value{
a data frame
}
\description{
This function loops through list of users and collects tweets between specified date ranges. Tweet-level data is stored in a data/ path as a series of JSONS; User-level data is stored in a includes/ path as a series of JSONS. If a filename is supplied, this function will save the result as a RDS file, otherwise, it will return the results as a dataframe.
}
\examples{
\dontrun{
bearer_token <- "XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX"
get_hashtag_tweets("#BLM", "2020-01-01T00:00:00Z", "2020-01-05T00:00:00Z", bearer_token, data_path = "data/")
}
}
