% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_liking_users.R
\name{get_liking_users}
\alias{get_liking_users}
\title{Get liking users}
\usage{
get_liking_users(x, bearer_token = get_bearer(), verbose = TRUE)
}
\arguments{
\item{x}{string containing one tweet id or a vector of tweet ids}

\item{bearer_token}{string, bearer token}

\item{verbose}{If \code{FALSE}, query progress messages are suppressed}
}
\value{
a data frame
}
\description{
This function fetches users who liked a tweet or tweets.
}
\examples{
\dontrun{
tweet <- "1387744422729748486"
get_liking_users(tweet, bearer_token = get_bearer())
}
}
