% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_media_tweetsv2.R
\name{get_media_tweets}
\alias{get_media_tweets}
\title{Get tweets for query containing media}
\usage{
get_media_tweets(
  query,
  start_tweets,
  end_tweets,
  bearer_token,
  file = NULL,
  data_path = NULL
)
}
\arguments{
\item{query}{}

\item{start_tweets}{}

\item{end_tweets}{}

\item{bearer_token}{}

\item{file}{}

\item{data_path}{}
}
\value{

}
\description{
This function loops through specified strings or hashtags and collects tweets containing the strings or hashtags between specified date ranges that also contain a media object, such as a photo, GIF, or video, as determined by Twitter. Tweet-level data is stored in a data/ path as a series of JSONs beginning "data_"; User-level data is stored as a series of JSONs beginning "user_". If a filename is supplied, this function will save the result as a RDS file, otherwise, it will return the results as a dataframe.
}
\examples{
\dontrun{
bearer_token <- "XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX"
get_media_tweets("#BLM", "2020-01-01T00:00:00Z", "2020-01-05T00:00:00Z", bearer_token, data_path = "data/")
}
}
