% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mentions_tweetsv2.R
\name{get_mentions_tweets}
\alias{get_mentions_tweets}
\title{Get tweets for query containing mentions of another user}
\usage{
get_mentions_tweets(
  query,
  start_tweets,
  end_tweets,
  bearer_token,
  file = NULL,
  data_path = NULL,
  bind_tweets = TRUE
)
}
\arguments{
\item{query}{string, search query, use "+" to separate query terms.}

\item{start_tweets}{string, starting date}

\item{end_tweets}{string, ending date}

\item{bearer_token}{string, bearer token}

\item{file}{string, name of the resulting RDS file}

\item{data_path}{string, if supplied, fetched data can be saved to the designated path as jsons}

\item{bind_tweets}{If \code{TRUE}, tweets captured are bound into a data.frame for assignment}
}
\value{

}
\description{
This function loops through specified strings or hashtags and collects tweets containing the strings or hashtags between specified date ranges that also contain mentions of another Twitter user. Tweet-level data is stored in a data/ path as a series of JSONs beginning "data_"; User-level data is stored as a series of JSONs beginning "users_". If a filename is supplied, this function will save the result as a RDS file, otherwise, it will return the results as a dataframe.
}
