% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_url_tweetsv2.R
\name{get_url_tweets}
\alias{get_url_tweets}
\title{Get tweets for query containing geo information}
\usage{
get_url_tweets(
  query,
  start_tweets,
  end_tweets,
  bearer_token,
  file = NULL,
  data_path = NULL
)
}
\arguments{
\item{query}{}

\item{start_tweets}{}

\item{end_tweets}{}

\item{bearer_token}{}

\item{file}{}

\item{data_path}{}
}
\value{

}
\description{
This function loops through specified urls to capture tweets containing the url between specified date ranges. Tweet-level data is stored in a data/ path as a series of JSONs beginning "data_"; User-level data is stored as a series of JSONs beginning "users_". If a filename is supplied, this function will save the result as a RDS file, otherwise, it will return the results as a dataframe.
}
