% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_user_id.R
\name{get_user_id}
\alias{get_user_id}
\title{Get user id}
\usage{
get_user_id(usernames, bearer_token = get_bearer(), all = FALSE)
}
\arguments{
\item{usernames}{string containing the screen name of the user}

\item{bearer_token}{string, bearer token}

\item{all}{logic, default FALSE. Set TRUE for getting also the screen name}
}
\value{
default a string vector with the id of each of the users.
If all = TRUE a data.frame with the id, name (showed on the screen) and username
}
\description{
This function get the user ID given a user name.
}
\examples{
\dontrun{
bearer_token <- "XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX"
users <- c("Twitter", "TwitterDev")
get_user_id(users, bearer_token)
}
}
