% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_user_profile.R
\name{get_user_profile}
\alias{get_user_profile}
\title{Get user profile}
\usage{
get_user_profile(x, bearer_token)
}
\arguments{
\item{x}{string containing one user id or a vector of user ids}

\item{bearer_token}{string, bearer token}
}
\value{
a data frame
}
\description{
This function fetch a variety of information about one or more users specified by the requested ids.
}
\examples{
\dontrun{
bearer_token <- "XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX"
users <- c("2244994945", "6253282")
get_user_profile(users, bearer_token)
}
}
