% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_user_status.R
\name{get_user_status}
\alias{get_user_status}
\title{Get user status}
\usage{
get_user_status(x, bearer_token = get_bearer())
}
\arguments{
\item{x}{string containing one user id or a vector of user ids}

\item{bearer_token}{string, bearer token}
}
\value{
a data frame returning five variables: id, verified, protected, suspended, undetectable. 'undetectable' is returned if a user ID never existed or once existed and is now deleted (not suspended but deleted, often by the user herself).
}
\description{
This function fetches information about the current status of a given user IDs (e.g. 123456789).
}
\examples{
\dontrun{
bearer_token <- "XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX"
users <- c("2244994945", "6253282")
get_user_status(users, bearer_token)
}
}
