% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compliance.R
\name{list_compliance_jobs}
\alias{list_compliance_jobs}
\title{List Compliance Jobs}
\usage{
list_compliance_jobs(type = "tweets", bearer_token = get_bearer())
}
\arguments{
\item{type}{the type of the job, whether "tweets" or "users".}

\item{bearer_token}{string, bearer token}
}
\value{
a data frame
}
\description{
This function lists all compliance jobs.
}
\examples{
\dontrun{
list_compliance_jobs()
}
}
