% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_bearer.R
\name{set_bearer}
\alias{set_bearer}
\alias{get_bearer}
\title{Manage your bearer token}
\usage{
set_bearer(bearer_token = NULL, path = "~/.academictwitteR_token")

get_bearer(path = "~/.academictwitteR_token")
}
\arguments{
\item{bearer_token}{string, your bearer token}

\item{path}{string, path to store your bearer token. Default to .academictwitteR_token at your user directory}
}
\value{
nothing. Your bearer token is stored
}
\description{
These two functions manage your bearer token. It is in general not safe to 1) hardcode your bearer token in your R script or 2) have your bearer token in your command history. \code{set_bearer} saves your bearer token as an RDS file. \code{get_bearer} returns your bearer token, if it has been preset.
}
