% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_collection.R
\name{update_collection}
\alias{update_collection}
\title{Update previous collection session}
\usage{
update_collection(
  data_path,
  end_tweets,
  bearer_token = get_bearer(),
  verbose = TRUE,
  ...
)
}
\arguments{
\item{data_path}{string, name of an existing data_path}

\item{end_tweets}{string, ending date}

\item{bearer_token}{string, bearer token}

\item{verbose}{If \code{FALSE}, query progress messages are suppressed}

\item{...}{arguments will be passed to \code{get_all_tweets()} function. See \code{?get_all_tweets()} for further information.}
}
\value{
a data.frame
}
\description{
This function continues a previous collection session with a new end date.
For this function to work, export_query must be set to "TRUE" during the original collection.
}
\examples{
\dontrun{
update_collection(data_path = "data", "2020-01-03T00:00:00Z", bearer_token = get_bearer())
}
}
