% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{bind_tweet_jsons}
\alias{bind_tweet_jsons}
\title{Bind tweets stored as JSON files
`r lifecycle::badge("deprecated")}
\usage{
bind_tweet_jsons(data_path, verbose = TRUE)
}
\arguments{
\item{data_path}{string, file path to directory of stored tweets data saved as data_\emph{id}.json}

\item{verbose}{If \code{FALSE}, messages are suppressed}
}
\value{
a data.frame
}
\description{
Bind tweets stored as JSON files
`r lifecycle::badge("deprecated")
}
\examples{
\dontrun{
bind_tweet_jsons(data_path = "data/")
}
}
\keyword{internal}
