% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_user_jsons.R
\name{bind_tweets}
\alias{bind_tweets}
\title{Bind information stored as JSON files}
\usage{
bind_tweets(data_path, user = FALSE, verbose = TRUE)
}
\arguments{
\item{data_path}{string, file path to directory of stored tweets data saved as data_\emph{id}.json and users_\emph{id}.json}

\item{user}{If \code{FALSE}, this function binds JSON files into a data frame containing tweets; data frame containing user information otherwise}

\item{verbose}{If \code{FALSE}, messages are suppressed}
}
\value{
a data.frame containing either tweets or user information
}
\description{
This function binds information stored as JSON files. By default, it binds into a data frame containing tweets (from data_\emph{id}.json files). If users is TRUE, it binds into a data frame containing user information (from users_\emph{id}.json).
}
\examples{
\dontrun{
# bind json files in the directory "data" into a data frame containing tweets
bind_tweets(data_path = "data/")
# bind json files in the directory "data" into a data frame containing user information
bind_tweets(data_path = "data/", user = TRUE)
}
}
