% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_user_jsons.R
\name{bind_user_jsons}
\alias{bind_user_jsons}
\title{Bind user information stored as JSON files}
\usage{
bind_user_jsons(data_path)
}
\arguments{
\item{data_path}{string, file path to directory of stored tweets data saved as users_\emph{id}.json}
}
\value{
a data.frame
}
\description{
Bind user information stored as JSON files
}
\examples{
\dontrun{
bind_user_jsons("data_path = "data/"")
}
}
