% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bearer.R
\name{get_bearer}
\alias{get_bearer}
\title{Manage bearer token}
\usage{
get_bearer()
}
\value{
string represents your bearer token, if it the environmental variable "TWITTER_BEARER" has been preset.
}
\description{
This function attempts to retrieve your bearer token from the environmental variable "TWITTER_BEARER".
The easiest way to setup this environmental variable is to use \code{set_bearer()}
and insert your bearer token to \code{.Renviron} file following the format: TWITTER_BEARER=YOURTOKENHERE.
Replace YOURTOKENHERE with your own token.
}
\details{
Note: for \code{get_bearer()} to retrieve your bearer token you will need to restart the
R session after storing in \code{.Renviron}.
}
