% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compliance.R
\name{get_compliance_result}
\alias{get_compliance_result}
\title{Get Compliance Result}
\usage{
get_compliance_result(id, bearer_token = get_bearer(), verbose = TRUE)
}
\arguments{
\item{id}{string, the job id}

\item{bearer_token}{string, bearer token}

\item{verbose}{If \code{FALSE}, query progress messages are suppressed}
}
\value{
a data frame
}
\description{
This function retrieves the information for a single compliance job.
}
\examples{
\dontrun{
get_compliance_result("1460077048991555585")
}
}
