% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resume_collection.R
\name{resume_collection}
\alias{resume_collection}
\title{Resume previous collection}
\usage{
resume_collection(data_path, bearer_token = get_bearer(), verbose = TRUE, ...)
}
\arguments{
\item{data_path}{string, name of an existing data_path}

\item{bearer_token}{string, bearer token}

\item{verbose}{If \code{FALSE}, query progress messages are suppressed}

\item{...}{arguments will be passed to \code{get_all_tweets()} function. See \code{?get_all_tweets()} for further information.}
}
\value{
a data.frame
}
\description{
This function resumes a previous interrupted collection session.
}
\details{
For this function to work, export_query must be set to "TRUE" during the original collection.
}
\examples{
\dontrun{
resume_collection(data_path = "data", bearer_token = get_bearer())
}
}
